refactor: rename jump_to_rem.ts to window.ts
The file only handles window object registration, so the simpler name is more appropriate given it's already in the register/ folder.
