+++ src/widgets/register/window.ts
